This is the final mid term submission.
I have made a desktop version. 
There is a section left and the footer is left, font style, font color, and the spacing is still not set in the portion done.
The position of the navigation Bar should be sticky.
The spacing in the navigation Bar is to be set, and the headings and paragraphs in section 1 are to be aligned and justifies in center.